(SIMP-5028) Disable usage ping by default
By default, every installation of GitLab sends a weekly report to GitLab
Inc, which contains usage details and the FQDN of the system. This
feature is opt-out, not opt-in.

This patch disables the update ping by default as an OPSEC safeguard,
consistent with NIST 800.53 rev 4 controls AC-20(1) and SC-38.

SIMP-5067 #close
SIMP-5028 #close