(SIMP-MAINT) Embed diagrams.net src into diagram
This patch re-implements the simp_gitlab components diagram image
(originally made with Gliffy as part of a Jira issue) in diagrams.net,
with the diagram source code embedded into the PNG file itself.

This will enable future contributors to load the image into the
diagrams.net editor and update the diagram.

This patch also corrects the component module name in the diagram from
`vshn-gitlan` to `puppet/gitlab`.